Add shared test-support headers and make fixtures cwd-independent
Introduces a header-only tests/support library (linked via a test_support
INTERFACE target) and routes fixture access through it:

- Fixtures.h resolves fixtures by absolute path via a GECK_TEST_DATA_DIR
  compile definition, now set for general_tests and performance_tests too
  (qt_tests already had it). Tests no longer depend on the working
  directory, so running a test binary directly or from an IDE/debugger no
  longer silently fails to find data. The redundant tests/data copy steps
  for general/performance are dropped.
- TempFile.h (RAII temp path + readAllBytes), ProStubProvider.h (StubProvider
  + pidOf), MapObjectBuilder.h (fillBase/checkBase) and ProBuilder.h
  (proRoundTrip) promote the helpers previously duplicated across the
  round-trip suites.

Both round-trip suites are migrated onto the shared helpers as the proof,
removing the per-test temp-path/cleanup boilerplate. All 118 tests pass.
