+++ tests/general/test_map_roundtrip.cpp
@@ -1,8 +1,8 @@
 #include <catch2/catch_test_macros.hpp>
 
+#include <cstddef>
 #include <cstdint>
 #include <filesystem>
-#include <map>
 #include <memory>
 
 #include "format/map/Map.h"
@@ -10,94 +10,12 @@
 #include "format/pro/Pro.h"
 #include "reader/map/MapReader.h"
 #include "writer/map/MapWriter.h"
+#include "support/MapObjectBuilder.h"
+#include "support/ProStubProvider.h"
+#include "support/TempFile.h"
 
 using namespace geck;
-
-namespace {
-
-// Stub PRO provider. Only ITEM/SCENERY objects dereference their Pro during
-// (de)serialization (for objectSubtypeId()); walls/critters/misc do not. So a
-// minimal in-memory Pro carrying just the subtype is enough, and the whole
-// round-trip needs no .map or .pro fixture on disk.
-struct StubProvider {
-    std::map<uint32_t, std::unique_ptr<Pro>> pros;
-
-    void addItem(uint32_t pid, Pro::ITEM_TYPE t) { set(pid, static_cast<unsigned int>(t)); }
-    void addScenery(uint32_t pid, Pro::SCENERY_TYPE t) { set(pid, static_cast<unsigned int>(t)); }
-
-    void set(uint32_t pid, unsigned int subtype) {
-        auto pro = std::make_unique<Pro>(std::filesystem::path("stub"));
-        pro->setObjectSubtypeId(subtype);
-        pros[pid] = std::move(pro);
-    }
-
-    Pro* load(uint32_t pid) const {
-        auto it = pros.find(pid);
-        return it != pros.end() ? it->second.get() : nullptr;
-    }
-};
-
-uint32_t pidOf(Pro::OBJECT_TYPE type, uint32_t index) {
-    return (static_cast<uint32_t>(type) << 24) | index;
-}
-
-std::filesystem::path tempMapPath() {
-    auto path = std::filesystem::temp_directory_path() / "geck_map_roundtrip.map";
-    std::error_code ec;
-    std::filesystem::remove(path, ec);
-    return path;
-}
-
-// Fills every serialized base field with a distinctive seed-derived value, so a
-// field-order or width bug anywhere in the 22-field block is caught for each
-// object type. pro_pid / elevation / inventory are set by the caller.
-void fillBase(MapObject& o, int32_t seed) {
-    o.unknown0 = static_cast<uint32_t>(seed * 100 + 1);
-    o.position = seed * 100 + 2;
-    o.x = static_cast<uint32_t>(seed * 100 + 3);
-    o.y = static_cast<uint32_t>(seed * 100 + 4);
-    o.sx = -(seed * 100 + 5);
-    o.sy = -(seed * 100 + 6);
-    o.frame_number = static_cast<uint32_t>(seed * 100 + 7);
-    o.direction = static_cast<uint32_t>(seed % 6);
-    o.frm_pid = static_cast<uint32_t>(seed * 100 + 9);
-    o.flags = static_cast<uint32_t>(seed * 100 + 10);
-    o.critter_index = seed * 100 + 11;
-    o.light_radius = static_cast<uint32_t>(seed % 9);
-    o.light_intensity = static_cast<uint32_t>(seed * 100 + 13);
-    o.outline_color = static_cast<uint32_t>(seed % 256);
-    o.map_scripts_pid = seed * 100 + 15;
-    o.script_id = seed * 100 + 16;
-    o.max_inventory_size = static_cast<uint32_t>(seed * 100 + 17);
-    o.unknown10 = static_cast<uint32_t>(seed * 100 + 18);
-    o.unknown11 = static_cast<uint32_t>(seed * 100 + 19);
-}
-
-void checkBase(const MapObject& got, const MapObject& want) {
-    CHECK(got.unknown0 == want.unknown0);
-    CHECK(got.position == want.position);
-    CHECK(got.x == want.x);
-    CHECK(got.y == want.y);
-    CHECK(got.sx == want.sx);
-    CHECK(got.sy == want.sy);
-    CHECK(got.frame_number == want.frame_number);
-    CHECK(got.direction == want.direction);
-    CHECK(got.frm_pid == want.frm_pid);
-    CHECK(got.flags == want.flags);
-    CHECK(got.elevation == want.elevation);
-    CHECK(got.pro_pid == want.pro_pid);
-    CHECK(got.critter_index == want.critter_index);
-    CHECK(got.light_radius == want.light_radius);
-    CHECK(got.light_intensity == want.light_intensity);
-    CHECK(got.outline_color == want.outline_color);
-    CHECK(got.map_scripts_pid == want.map_scripts_pid);
-    CHECK(got.script_id == want.script_id);
-    CHECK(got.max_inventory_size == want.max_inventory_size);
-    CHECK(got.unknown10 == want.unknown10);
-    CHECK(got.unknown11 == want.unknown11);
-}
-
-} // namespace
+using namespace geck::test;
 
 // MapWriter/MapReader symmetry guard. Builds one object of every type and item/
 // scenery subtype that has a distinct serialization path (plus a nested inventory
@@ -181,7 +99,8 @@ TEST_CASE("MAP round-trip preserves all object types and inventory", "[map][roun
 
     const size_t objectCount = objects.size();
 
-    const auto path = tempMapPath();
+    TempFile mapFile{ "geck_map_roundtrip_alltypes", ".map" };
+    const auto& path = mapFile.path();
     {
         MapWriter writer{ [&](int32_t pid) { return provider.load(static_cast<uint32_t>(pid)); } };
         writer.openFile(path);
@@ -248,9 +167,6 @@ TEST_CASE("MAP round-trip preserves all object types and inventory", "[map][roun
     CHECK(gotExit.exit_position == exitGrid->exit_position);
     CHECK(gotExit.exit_elevation == exitGrid->exit_elevation);
     CHECK(gotExit.exit_orientation == exitGrid->exit_orientation);
-
-    std::error_code ec;
-    std::filesystem::remove(path, ec);
 }
 
 // MapObject::cloneDeep underpins the inventory and copy-elevation undo commands,
@@ -315,7 +231,8 @@ TEST_CASE("MAP single-enabled-elevation map keeps engine 3-block object framing"
 
     const size_t objectCount = objects.size();
 
-    const auto path = tempMapPath();
+    TempFile mapFile{ "geck_map_roundtrip_single_elev", ".map" };
+    const auto& path = mapFile.path();
     {
         MapWriter writer{ [&](int32_t pid) { return provider.load(static_cast<uint32_t>(pid)); } };
         writer.openFile(path);
@@ -340,9 +257,6 @@ TEST_CASE("MAP single-enabled-elevation map keeps engine 3-block object framing"
     CHECK(result.tiles.count(0) == 1);
     CHECK(result.tiles.count(1) == 0);
     CHECK(result.tiles.count(2) == 0);
-
-    std::error_code ec;
-    std::filesystem::remove(path, ec);
 }
 
 // A non-exit-grid MISC object must not serialize the 4 trailing exit fields:
@@ -378,11 +292,10 @@ TEST_CASE("MAP non-exit MISC object writes no trailing exit data", "[map][roundt
         return reader.openFile(p);
     };
 
-    const auto exitPath = std::filesystem::temp_directory_path() / "geck_misc_exit.map";
-    const auto plainPath = std::filesystem::temp_directory_path() / "geck_misc_plain.map";
-    std::error_code ec;
-    std::filesystem::remove(exitPath, ec);
-    std::filesystem::remove(plainPath, ec);
+    TempFile exitFile{ "geck_misc_exit", ".map" };
+    TempFile plainFile{ "geck_misc_plain", ".map" };
+    const auto& exitPath = exitFile.path();
+    const auto& plainPath = plainFile.path();
 
     writeMap(buildSingleMisc(16), exitPath); // exit grid -> 4 trailing fields
     writeMap(buildSingleMisc(5), plainPath); // ordinary MISC -> none
@@ -413,7 +326,4 @@ TEST_CASE("MAP non-exit MISC object writes no trailing exit data", "[map][roundt
     CHECK(exitObj.exit_position == 902);
     CHECK(exitObj.exit_elevation == 903);
     CHECK(exitObj.exit_orientation == 904);
-
-    std::filesystem::remove(exitPath, ec);
-    std::filesystem::remove(plainPath, ec);
 }
